[Bugfix][Model Runner V2] Cast idx_mapping to int64 for index_fill_ in mamba hybrid postprocess_state
The scalar branch of MambaHybridModelState.postprocess_state calls
index_fill_ with idx_mapping, which is int32. index_fill_ accepts only
int64 indices, so the call raises:

    IndexError: index_fill_(): Expected dtype int64 for index.

This branch has exactly one caller: the non-last-PP-rank path in
GPUModelRunner.execute_model, which passes a literal 0 for num_sampled.
Reaching it therefore requires a hybrid (Mamba/KDA) model running with
pipeline_parallel_size >= 2 and a batch containing a non-final prefill
chunk, which is why the dtype error has gone unnoticed.

The mapping passed at that call site is input_batch.idx_mapping, which is
dense and contains no -1 sentinels, so widening it is safe.

+++ vllm/v1/worker/gpu/model_states/mamba_hybrid.py
@@ -303,9 +303,12 @@ class MambaHybridModelState(DefaultModelState):
                     idx_mapping, num_sampled, self.num_accepted_tokens_gpu
                 )
         else:
-            # Fill with single value.
+            # Fill with single value. index_fill_ only accepts int64 indices,
+            # while idx_mapping is int32. This branch is reached only from the
+            # non-last PP rank, which passes a dense mapping with no -1
+            # sentinels, so the cast is safe.
             self.num_accepted_tokens_gpu.index_fill_(
-                0, idx_mapping, max(num_sampled, 1)
+                0, idx_mapping.long(), max(num_sampled, 1)
             )
 
         # Align: save the running state to the block-aligned position when
